Выполнены рекомендуемые доработки после ревью: 1. Из класса WebViewController удалены методы viewWillAppear и viewDidDisappear. 2. @IBAction в классах ImageListCell и ProfileViewController теперь находится после приватных функций. 3. У метода performBatchUpdate (класс ImageListViewController) удален completion: { _ in }. 4. У метода updateProfileDetails (класс ProfileViewController) удален модификатор доступа internal (так как он установлен по умолчанию). 5. Классы WebViewControllerSpy и WebViewPresenterSpy выделены в отдельные файлы. 6. Переменные emailLogin и password перемещены в поля класса ImageFeedUITests. Там же объявлены новые переменные nameLabel и loginNameLabel. 7. Файл ImageFeedUITestsLaunchTests.swift удален.
