+++ ImageFeed/ImagesList/ImagesListCell.swift
@@ -14,10 +14,6 @@ final class ImagesListCell: UITableViewCell {
     
     weak var delegate: ImagesListCellDelegate?
     
-    @IBAction private func likeButtonClicked(_ sender: Any) {
-        delegate?.imageListCellDidTapLike(self)
-    }
-    
     static let reuseIdentifier = "ImagesListCell"
     
     private enum LikePhoto: String {
@@ -33,5 +29,9 @@ final class ImagesListCell: UITableViewCell {
     func setIsLiked(isLiked: Bool) {
         let likePhoto = isLiked ? UIImage(named: "like_active") : UIImage(named: "like_inactive")
         likeButton.setImage(likePhoto, for: .normal)
+    }
+    
+    @IBAction private func likeButtonClicked(_ sender: Any) {
+        delegate?.imageListCellDidTapLike(self)
     }
 }
