+++ ImageFeedTest/WebViewControllerSpy.swift
@@ -1,0 +1,27 @@
+//
+//  WebViewControllerSpy.swift
+//  ImageFeedTest
+//
+//  Created by Владимир Богомолов on 22.08.2023.
+//
+
+import ImageFeed
+import Foundation
+
+final class WebViewViewControllerSpy: WebViewViewControllerProtocol {
+    var presenter: ImageFeed.WebViewPresenterProtocol?
+    
+    var loadRequestCalled: Bool = false
+    
+    func load(request: URLRequest) {
+        loadRequestCalled = true
+    }
+    
+    func setProgressValue(_ newValue: Float) {
+        
+    }
+    
+    func setProgressHidden(_ isHidden: Bool) {
+        
+    }
+}
