+++ ImageFeedTest/WebViewPresenterSpy.swift
@@ -1,0 +1,26 @@
+//
+//  WebViewPresenterSpy.swift
+//  ImageFeedTest
+//
+//  Created by Владимир Богомолов on 22.08.2023.
+//
+
+import ImageFeed
+import Foundation
+
+final class WebViewPresenterSpy: WebViewPresenterProtocol {
+    var viewDidLoadCalled: Bool = false
+    var view: WebViewViewControllerProtocol?
+    
+    func viewDidLoad() {
+        viewDidLoadCalled = true
+    }
+    
+    func didUpdateProgressValue(_ newValue: Double) {
+        
+    }
+    
+    func code(from url: URL) -> String? {
+        return nil
+    }
+}
