+++ ImageFeedTest/WebViewTests.swift
@@ -9,41 +9,6 @@
 
 import XCTest
 
-final class WebViewViewControllerSpy: WebViewViewControllerProtocol {
-    var presenter: ImageFeed.WebViewPresenterProtocol?
-    
-    var loadRequestCalled: Bool = false
-    
-    func load(request: URLRequest) {
-        loadRequestCalled = true
-    }
-    
-    func setProgressValue(_ newValue: Float) {
-        
-    }
-    
-    func setProgressHidden(_ isHidden: Bool) {
-        
-    }
-}
-
-final class WebViewPresenterSpy: WebViewPresenterProtocol {
-    var viewDidLoadCalled: Bool = false
-    var view: WebViewViewControllerProtocol?
-    
-    func viewDidLoad() {
-        viewDidLoadCalled = true
-    }
-    
-    func didUpdateProgressValue(_ newValue: Double) {
-        
-    }
-    
-    func code(from url: URL) -> String? {
-        return nil
-    }
-}
-
 final class ImageFeedTest: XCTestCase {
     func testViewControllerCallsViewDidLoad() {
         //given
